docs: improve docs
