+++ docs/package.json
@@ -22,7 +22,6 @@
     "@types/node": "^18.15.3",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
-    "next-transpile-modules": "^10.0.0",
     "typescript": "^5.0.2"
   }
 }
